requirements.txt
flask==1.1.2
requests==2.25.1
pandas==1.2.4
psycopg2-binary==2.8.6
